1. bilibili视频 添加只下载音频的场景

+++ biliVideo/run.py
@@ -16,22 +16,27 @@ from common.utils import get_common_headers, NetworkException
 
 
 class MergeThread(threading.Thread):
-    def __init__(self, path, filename):
+    def __init__(self, path, filename, done=False):
         self.path = path
         self.filename = filename
+        self.done = done
         super().__init__()
 
     def run(self):
-        print('开始合并视频音频...')
-        ad = AudioFileClip(f'{self.path}.mp3')
-        vd = VideoFileClip(f'{self.path}.mp4')
-        vd.set_audio(ad).write_videofile(f'{self.filename}.mp4', logger=None)
-        print(f'【{self.filename}】合并完成.')
-        os.remove(f'{self.path}.mp3')
-        os.remove(f'{self.path}.mp4')
+        if not self.done:
+            print('开始合并视频音频...')
+            ad = AudioFileClip(f'{self.path}.mp3')
+            vd = VideoFileClip(f'{self.path}.mp4')
+            vd.set_audio(ad).write_videofile(f'{self.filename}.mp4', logger=None)
+            print(f'【{self.filename}】合并完成.')
+            os.remove(f'{self.path}.mp3')
+            os.remove(f'{self.path}.mp4')
+        else:
+            if os.path.exists(f'{self.path}.mp3'):
+                os.rename(f'{self.path}.mp3', f'{self.filename}.mp3')
 
 
-def run(bvid):
+def run(bvid, only_audio=False):
     url = f'https://www.bilibili.com/video/{bvid}'
     headers = {
         'referer': 'https://www.bilibili.com',
@@ -61,10 +66,11 @@ def run(bvid):
     audio_url = dash['audio'][0]['baseUrl']
 
     try:
-        print('视频音频下载开始...')
-        video_content = requests.get(video_url, headers=headers).content
+        print('视频音视频下载开始...')
+        if not only_audio:
+            video_content = requests.get(video_url, headers=headers).content
         audio_content = requests.get(audio_url, headers=headers).content
-        print('视频音频下载完成.')
+        print('视频音视频下载完成.')
     except Exception:
         raise Exception
 
@@ -74,19 +80,19 @@ def run(bvid):
             shutil.rmtree(path)
         os.mkdir(path)
 
-        with open(path + video_name + '.mp4', 'wb') as f:
-            f.write(video_content)
         with open(path + video_name + '.mp3', 'wb') as f:
             f.write(audio_content)
-        return MergeThread(path + video_name, f'{path}【{up_name}】 {video_name}')
+        if not only_audio:
+            with open(path + video_name + '.mp4', 'wb') as f:
+                f.write(video_content)
+
+        return MergeThread(path + video_name, f'{path}【{up_name}】 {video_name}', done=only_audio)
     except Exception:
         raise Exception
 
 
 if __name__ == '__main__':
     bvid = input('请输入视频Bv号:')
-    while len(bvid) > 0:
-        thread = run(bvid)
-        thread.start()
-        bvid = input('\n请输入视频Bv号:')
+    thread = run(bvid, True)
+    thread.start()
     print('Exit.')
